requirements.txt
asgiref==3.4.1
Babel==2.9.1
Django==3.2
django-jazzmin==2.4.9
django-money==2.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
Pillow==5.0.0
pkg-resources==0.0.0
py-moneyed==1.2
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
